fix(scan): clear first-party CodeQL integer-widening warnings (#1698)
Cast first factor to size_t in vector-size, index, and memcpy arithmetic
(cpp/integer-multiplication-cast-to-long) across src/ engine files.
Escape created_at/expires_at in dashboard API-key table (js/xss-through-dom).
Cast same to size_t in printf %zu (cpp/wrong-type-format-argument).

Vendored third_party alerts (1,013) dismissed upstream as won't-fix.

+++ src/afmoe_engine.cpp
@@ -266,8 +266,8 @@ private:
     }
 
     void attention_mix(const float* x, const AfmoeLayer& ly, int l, int pos, float* out) {
-        std::vector<float> q(NH * HD), k(NKV * HD), v(NKV * HD), gate(NH * HD);
-        std::vector<float> qn(NH * HD), kn(NKV * HD), qr(NH * HD), kr(NKV * HD);
+        std::vector<float> q((size_t)NH * HD), k((size_t)NKV * HD), v((size_t)NKV * HD), gate((size_t)NH * HD);
+        std::vector<float> qn((size_t)NH * HD), kn((size_t)NKV * HD), qr((size_t)NH * HD), kr((size_t)NKV * HD);
         mm(ly.q_proj, x, H, NH * HD, q.data());
         mm(ly.k_proj, x, H, NKV * HD, k.data());
         mm(ly.v_proj, x, H, NKV * HD, v.data());
@@ -278,22 +278,22 @@ private:
             rope_apply(qn.data(), NH, HD, qr.data(), pos);
             rope_apply(kn.data(), NKV, HD, kr.data(), pos);
         } else {
-            memcpy(qr.data(), qn.data(), NH * HD * sizeof(float));
-            memcpy(kr.data(), kn.data(), NKV * HD * sizeof(float));
+            memcpy(qr.data(), qn.data(), (size_t)NH * HD * sizeof(float));
+            memcpy(kr.data(), kn.data(), (size_t)NKV * HD * sizeof(float));
         }
         auto& kk = attn_k[l]; auto& vv = attn_v[l];
         for (int i = 0; i < NKV * HD; i++) { kk.push_back(kr[i]); vv.push_back(v[i]); }
         int T = (int)kk.size() / (NKV * HD);
         float scale = 1.0f / std::sqrt((float)HD);
         int groups = NH / NKV;
-        std::vector<float> out_heads(NH * HD);
+        std::vector<float> out_heads((size_t)NH * HD);
         for (int hh = 0; hh < NH; hh++) {
             int kh = hh / groups;
             std::vector<float> scores(T);
             float mx = -1e30f;
             for (int t = 0; t < T; t++) {
                 float s = 0;
-                for (int d = 0; d < HD; d++) s += qr[hh * HD + d] * kk[(size_t)t * NKV * HD + kh * HD + d];
+                for (int d = 0; d < HD; d++) s += qr[(size_t)hh * HD + d] * kk[(size_t)t * NKV * HD + (size_t)kh * HD + d];
                 scores[t] = s * scale;
                 if (scores[t] > mx) mx = scores[t];
             }
@@ -301,7 +301,7 @@ private:
             for (int t = 0; t < T; t++) { scores[t] = std::exp(scores[t] - mx); sum += scores[t]; }
             for (int d = 0; d < HD; d++) {
                 float acc = 0;
-                for (int t = 0; t < T; t++) acc += scores[t] / sum * vv[(size_t)t * NKV * HD + kh * HD + d];
+                for (int t = 0; t < T; t++) acc += scores[t] / sum * vv[(size_t)t * NKV * HD + (size_t)kh * HD + d];
                 out_heads[hh * HD + d] = acc;
             }
         }
